requirements.txt
pandas
sqlalchemy
pymysql
prefect>=2.16
python-dotenv
fastapi
uvicorn
